# src/main/gen/java/com/xyz/shows/management/api/TheatersApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (3.0.15).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package com.xyz.shows.management.api;

import com.xyz.shows.management.model.CreateShowRequest;
import com.xyz.shows.management.model.CreateTheaterRequest;
import com.xyz.shows.management.model.Error;
import com.xyz.shows.management.model.Theater;
import com.fasterxml.jackson.databind.ObjectMapper;
import io.swagger.annotations.*;
import org.slf4j.Logger;
import org.slf4j.LoggerFactory;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestHeader;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RequestPart;
import org.springframework.web.multipart.MultipartFile;
import org.springframework.web.bind.annotation.CookieValue;

import javax.servlet.http.HttpServletRequest;
import javax.validation.Valid;
import javax.validation.constraints.*;
import java.io.IOException;
import java.util.List;
import java.util.Map;
import java.util.Optional;
@Api(value = "theaters", description = "the theaters API")
public interface TheatersApi {

    Logger log = LoggerFactory.getLogger(TheatersApi.class);

    default Optional<ObjectMapper> getObjectMapper(){
        return Optional.empty();
    }

    default Optional<HttpServletRequest> getRequest(){
        return Optional.empty();
    }

    default Optional<String> getAcceptHeader() {
        return getRequest().map(r -> r.getHeader("Accept"));
    }

    @ApiOperation(value = "Add a new theater", nickname = "addNewTheater", notes = "", response = Theater.class, tags={ "Theaters", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "Excepted response to a valid request", response = Theater.class),
        @ApiResponse(code = 409, message = "Theater already present in the inventory", response = Error.class),
        @ApiResponse(code = 400, message = "Incorrect request", response = Error.class),
        @ApiResponse(code = 500, message = "Internal server error", response = Error.class) })
    @RequestMapping(value = "/theaters",
        produces = { "application/json" }, 
        consumes = { "application/json" },
        method = RequestMethod.POST)
    default ResponseEntity<Theater> addNewTheater(@ApiParam(value = "Request body for adding a new theather to the inventory"  )  @Valid @RequestBody CreateTheaterRequest body
) {
        if(getObjectMapper().isPresent() && getAcceptHeader().isPresent()) {
            if (getAcceptHeader().get().contains("application/json")) {
                try {
                    return new ResponseEntity<>(getObjectMapper().get().readValue("{\r\n  \"zip\" : \"560001\",\r\n  \"address2\" : \"R Block\",\r\n  \"city\" : \"Bangalore\",\r\n  \"phone\" : \"12245-67890\",\r\n  \"screens\" : [ {\r\n    \"seatCapacity\" : 200,\r\n    \"name\" : \"Screen_1\"\r\n  }, {\r\n    \"seatCapacity\" : 200,\r\n    \"name\" : \"Screen_1\"\r\n  } ],\r\n  \"address1\" : \"100 abc road\",\r\n  \"name\" : \"PVR Cinemas\",\r\n  \"description\" : \"PVR Cinemas in bangalore\",\r\n  \"id\" : 501,\r\n  \"state\" : \"Karnataka\",\r\n  \"category\" : \"Multiplex\"\r\n}", Theater.class), HttpStatus.NOT_IMPLEMENTED);
                } catch (IOException e) {
                    log.error("Couldn't serialize response for content type application/json", e);
                    return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
                }
            }
        } else {
            log.warn("ObjectMapper or HttpServletRequest not configured in default TheatersApi interface so no example is generated");
        }
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);
    }


    @ApiOperation(value = "Get Theater details for a theaterId", nickname = "getTheaterDetails", notes = "", response = Theater.class, tags={ "Theaters", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "Success output", response = Theater.class),
        @ApiResponse(code = 404, message = "Cinema not found in inventory", response = Error.class),
        @ApiResponse(code = 400, message = "Incorrect request", response = Error.class),
        @ApiResponse(code = 500, message = "Internal server error", response = Error.class) })
    @RequestMapping(value = "/theaters/{theaterId}",
        produces = { "application/json" }, 
        method = RequestMethod.GET)
    default ResponseEntity<Theater> getTheaterDetails(@ApiParam(value = "Id of a Theater",required=true) @PathVariable("theaterId") Long theaterId
) {
        if(getObjectMapper().isPresent() && getAcceptHeader().isPresent()) {
            if (getAcceptHeader().get().contains("application/json")) {
                try {
                    return new ResponseEntity<>(getObjectMapper().get().readValue("{\r\n  \"zip\" : \"560001\",\r\n  \"address2\" : \"R Block\",\r\n  \"city\" : \"Bangalore\",\r\n  \"phone\" : \"12245-67890\",\r\n  \"screens\" : [ {\r\n    \"seatCapacity\" : 200,\r\n    \"name\" : \"Screen_1\"\r\n  }, {\r\n    \"seatCapacity\" : 200,\r\n    \"name\" : \"Screen_1\"\r\n  } ],\r\n  \"address1\" : \"100 abc road\",\r\n  \"name\" : \"PVR Cinemas\",\r\n  \"description\" : \"PVR Cinemas in bangalore\",\r\n  \"id\" : 501,\r\n  \"state\" : \"Karnataka\",\r\n  \"category\" : \"Multiplex\"\r\n}", Theater.class), HttpStatus.NOT_IMPLEMENTED);
                } catch (IOException e) {
                    log.error("Couldn't serialize response for content type application/json", e);
                    return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
                }
            }
        } else {
            log.warn("ObjectMapper or HttpServletRequest not configured in default TheatersApi interface so no example is generated");
        }
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);
    }


    @ApiOperation(value = "Update Theater details for a cinemaId", nickname = "updateTheaterDetails", notes = "", response = Theater.class, tags={ "Theaters", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "Success output", response = Theater.class),
        @ApiResponse(code = 404, message = "Cinema not found in inventory", response = Error.class),
        @ApiResponse(code = 400, message = "Incorrect request", response = Error.class),
        @ApiResponse(code = 500, message = "Internal server error", response = Error.class) })
    @RequestMapping(value = "/theaters/{theaterId}",
        produces = { "application/json" }, 
        consumes = { "application/json" },
        method = RequestMethod.PUT)
    default ResponseEntity<Theater> updateTheaterDetails(@ApiParam(value = "Id of a Theater",required=true) @PathVariable("theaterId") Long theaterId
,@ApiParam(value = "Show that needs to be added to the inventory"  )  @Valid @RequestBody CreateShowRequest body
) {
        if(getObjectMapper().isPresent() && getAcceptHeader().isPresent()) {
            if (getAcceptHeader().get().contains("application/json")) {
                try {
                    return new ResponseEntity<>(getObjectMapper().get().readValue("{\r\n  \"zip\" : \"560001\",\r\n  \"address2\" : \"R Block\",\r\n  \"city\" : \"Bangalore\",\r\n  \"phone\" : \"12245-67890\",\r\n  \"screens\" : [ {\r\n    \"seatCapacity\" : 200,\r\n    \"name\" : \"Screen_1\"\r\n  }, {\r\n    \"seatCapacity\" : 200,\r\n    \"name\" : \"Screen_1\"\r\n  } ],\r\n  \"address1\" : \"100 abc road\",\r\n  \"name\" : \"PVR Cinemas\",\r\n  \"description\" : \"PVR Cinemas in bangalore\",\r\n  \"id\" : 501,\r\n  \"state\" : \"Karnataka\",\r\n  \"category\" : \"Multiplex\"\r\n}", Theater.class), HttpStatus.NOT_IMPLEMENTED);
                } catch (IOException e) {
                    log.error("Couldn't serialize response for content type application/json", e);
                    return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
                }
            }
        } else {
            log.warn("ObjectMapper or HttpServletRequest not configured in default TheatersApi interface so no example is generated");
        }
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);
    }

}